+++ src/commands/sbom/constants.ts
@@ -1,1 +1,2 @@
+export const API_ENDPOINT = 'api/v2/sbom'
 export const INTAKE_NAME = 'sbom-intake'
